requirements.txt
https://github.com/wazo-platform/wazo-lib-rest-client/archive/master.zip
https://github.com/wazo-platform/wazo-agentd-client/archive/master.zip
https://github.com/wazo-platform/wazo-auth-client/archive/master.zip
https://github.com/wazo-platform/wazo-calld-client/archive/master.zip
https://github.com/wazo-platform/wazo-confd-client/archive/master.zip
https://github.com/wazo-platform/xivo-dao/archive/master.zip
https://github.com/wazo-platform/wazo-dird-client/archive/master.zip
https://github.com/wazo-platform/xivo-lib-python/archive/master.zip
psycopg2-binary==2.8.6
pytz==2021.1
pyyaml==5.3.1  # from xivo-lib-python
requests==2.25.1
sqlalchemy==1.3.22  # from xivo-dao
sqlalchemy-utils==0.36.8  # from xivo-dao
stevedore==3.2.2  # from wazo-lib-rest-client
unidecode==1.2.0
phonenumbers==8.12.1
